requirements.txt
actionlib==1.14.0
angles==1.9.13
attrs==19.3.0
Automat==0.8.0
av==12.3.0
bcrypt==3.1.7
blinker==1.4
breezy==3.0.2
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
catkin-tools==0.9.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==23.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.12.1
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
distlib==0.3.9
distro==1.4.0
distro-info===0.23ubuntu1
djitellopy==2.5.0
docutils==0.16
dulwich==0.19.15
empy==3.3.2
entrypoints==0.3
fastimport==0.9.8
filelock==3.16.1
fonttools==4.55.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gpg==1.13.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib-metadata==1.5.0
importlib-resources==6.4.5
incremental==16.10.1
interactive-markers==1.12.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupSafe==1.1.0
matplotlib==3.7.5
message-filters==1.17.0
more-itertools==4.2.0
netifaces==0.10.4
nose==1.3.7
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.10.0.84
osrf-pycommon==2.0.2
packaging==24.2
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==4.3.6
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
pygame==2.6.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.4
pytz==2019.3
PyYAML==5.3.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
requests==2.22.0
requests-unixsocket==0.2.0
resource-retriever==1.12.7
roman==2.0.0
rosbag==1.17.0
rosclean==1.15.8
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph==1.17.0
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.5.1
rospkg-modules==1.5.1
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rviz==1.14.25
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sos==4.4
ssh-import-id==5.10
systemd-python==234
tellopy==0.6.0
tf==1.13.2
tf2-geometry-msgs==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
topic-tools==1.17.0
Twisted==18.9.0
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vcstools==0.1.42
virtualenv==20.27.1
wadllib==1.3.3
wrapt==1.11.2
wstool==0.1.18
zipp==3.20.2
zope.interface==4.7.1
